python/level1_single_api/9_amct/amct_tensorflow/requirements.txt
matplotlib
numpy
Pillow
tensorflow == 1.15
tqdm
